Rename posterior time series heatmap to posterior visualizer
